Add exponential-backoff Update to use vite
