+++ .eslintignore
@@ -1,1 +1,2 @@
 types/*
+vite.config.ts
