+++ demo/test.html
@@ -30,9 +30,9 @@
 </head>
 <body>
 
-<pub-socket chat-fid="FID:INT:1670595842:IJd5Lb601tZVM"
-            chat-ref="INT-rmmo02-IJd5Lb601tZVM-NZETnXN3Py9PQvFTHbAO4vCm7hObAxQP"
-            socket-host="ws://socket.fortifi.me:8012"></pub-socket>
+<pub-socket chat-fid="FID:INT:1725539569:giEAhvYxPBAA7"
+            chat-ref="INT-sjcaup-giEAhvYxPBAA7-cCVWSaMrm1AczV7KIcH9j1rn4M0qDyWd"
+            socket-host="wss://socket.fortifi.me"></pub-socket>
 
 </body>
 </html>
