+++ package.json
@@ -5,32 +5,23 @@
   "main": "src/pubsocket.ts",
   "version": "1.0.0",
   "scripts": {
-    "build": "rollup -c rollup.config.js"
+    "build": "vite build",
+    "watch": "vite build --watch"
   },
   "dependencies": {
-    "lit": "^2.5.0",
-    "smile2emoji": "^3.5.0",
+    "lit": "^3.2.0",
+    "smile2emoji": "^3.14.0",
     "ulid": "^2.3.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.17.10",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/plugin-proposal-decorators": "^7.20.5",
-    "@babel/preset-env": "^7.17.10",
-    "@babel/preset-typescript": "^7.16.7",
-    "@rollup/plugin-babel": "^5.3.1",
-    "@rollup/plugin-commonjs": "^22.0.0",
-    "@rollup/plugin-eslint": "^8.0.2",
-    "@rollup/plugin-node-resolve": "^13.3.0",
     "@typescript-eslint/eslint-plugin": "^5.22.0",
     "@typescript-eslint/parser": "^5.22.0",
     "browserslist": "^4.20.3",
-    "core-js": "3",
     "eslint": "^8.15.0",
-    "rollup": "^2.72.1",
-    "rollup-plugin-command": "^1.1.3",
-    "rollup-plugin-polyfill": "^4.0.0",
-    "rollup-plugin-svg": "^2.0.0",
-    "rollup-plugin-terser": "^7.0.2"
+    "rollup-plugin-minify-html-literals": "^1.2.6",
+    "terser": "^5.31.6",
+    "typescript": "^5.5.4",
+    "vite": "^5.4.3",
+    "vitest": "^2.0.5"
   }
 }
