+++ src/pubsocket.ts
@@ -1,8 +1,10 @@
-import { css, html, LitElement, PropertyValues } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
-import { detectPrng, factory } from "ulid";
-import { unsafeHTML } from 'lit/directives/unsafe-html.js'
-import { checkText } from 'smile2emoji'
+import {CSSResultGroup, html, LitElement, PropertyValues} from 'lit';
+import {customElement, property} from 'lit/decorators.js';
+import {detectPrng, factory} from "ulid";
+import {unsafeHTML} from 'lit/directives/unsafe-html.js'
+import {checkText} from 'smile2emoji'
+
+import styles from './pubsocket.styles';
 
 const ulid = factory(detectPrng(true));
 
@@ -31,174 +33,42 @@ declare type Message = {
 }
 
 @customElement('pub-socket')
-export class PubSocket extends LitElement // eslint-disable-line @typescript-eslint/no-unused-vars
-{
+export class PubSocket extends LitElement {
+
+  static styles: CSSResultGroup = [styles]
+
   private _socket: WebSocket | null;
   private _lastMessageTime: number = 0;
   private _failureTime: number = 0;
   protected _messages: Message[] = [];
   protected _agentTyping = false;
-
   public scrolled: boolean = false;
+  protected _retryAttempts: number = 0;
 
-  @property({ attribute: 'new-message', reflect: true, type: Boolean })
+  @property({attribute: 'new-message', reflect: true, type: Boolean})
   public newMessage: boolean = false;
-  @property({ attribute: 'connected', reflect: true, type: Boolean })
+  @property({attribute: 'connected', reflect: true, type: Boolean})
   public connected: boolean = false;
-  @property({ attribute: 'connection-failed', reflect: true, type: Boolean })
+  @property({attribute: 'connection-failed', reflect: true, type: Boolean})
   public connectionFailed: boolean = false;
 
-  @property({ attribute: 'hide-send-panel', type: Boolean })
+  @property({attribute: 'hide-send-panel', type: Boolean})
   public hideSendPanel: boolean = false;
 
-  @property({ attribute: 'retry-delay', type: Number })
+  @property({attribute: 'retry-delay', type: Number})
   public retryDelay: number = 2;
-  @property({ attribute: 'retry-timeout', type: Number })
+  @property({attribute: 'retry-timeout', type: Number})
   public retryTimeout: number = 5;
 
-  @property({ attribute: 'socket-host' })
+  @property({attribute: 'socket-host'})
   public socketHost: string = 'wss://socket.fortifi.io';
 
-  @property({ attribute: 'chat-fid' })
+  @property({attribute: 'chat-fid'})
   public chatFid: string = '';
-  @property({ attribute: 'chat-ref' })
+  @property({attribute: 'chat-ref'})
   public chatRef: string = '';
 
   public canReply: boolean = true;
-
-  static styles = css`
-    :host {
-      --customer-background-color: #ccf5db;
-      position: relative;
-      display: flex;
-      flex-direction: column;
-      width: 100%;
-      height: 100%;
-      overflow: auto;
-    }
-
-    ul {
-      padding: 0;
-      display: flex;
-      flex-direction: column;
-      width: var(--message-container-width, auto);
-      margin: var(--message-container-margin, 0);
-    }
-
-    li {
-      color: #000;
-      border-radius: 5px;
-      border: 1px solid rgba(0, 0, 0, .1);
-      background: var(--agent-background-color, #daf5ff);
-      padding: 7px;
-      margin: 4px;
-      display: flex;
-      width: fit-content;
-      flex-wrap: wrap;
-      align-self: flex-start;
-    }
-
-    li[action-type="connected.agent"],
-    li[action-type="attachment.added"] {
-      background: transparent;
-      border: none;
-      align-self: center;
-    }
-
-    li[action-type="connected.agent"] > span,
-    li[action-type="attachment.added"] > span {
-      display: none
-    }
-
-    li[customer] {
-      background: var(--customer-background-color);
-      align-self: flex-end;
-    }
-
-    li[undelivered] {
-      background: lightgray;
-    }
-
-    li .int-msg-who {
-      opacity: .7;
-      font-size: .7em;
-      padding: 0 0 3px 0;
-      align-self: start;
-      flex-basis: 100%;
-      font-weight: 700;
-    }
-
-    li .int-msg-time {
-      opacity: .6;
-      font-size: .7em;
-      padding: 5px 0 0 5px;
-      align-self: end;
-      flex-grow: 1;
-      text-align: right;
-    }
-
-    input, button {
-      padding: 5px;
-      margin: 5px;
-    }
-
-    #msg {
-      flex-grow: 1;
-    }
-
-    #send-panel {
-      display: flex;
-      position: sticky;
-      bottom: 0;
-      width: 100%;
-      background: white;
-    }
-
-    #agent-typing {
-      width: 100%;
-      margin: 0 auto;
-      font-size: 13px;
-    }
-
-    #agent-typing:after {
-      overflow: hidden;
-      display: inline-block;
-      vertical-align: bottom;
-      -webkit-animation: ellipsis steps(4, end) 1500ms infinite;
-      animation: ellipsis steps(4, end) 1500ms infinite;
-      content: "\\2026";
-      width: 0;
-    }
-
-    .int-msg-answers {
-      align-self: flex-end;
-      flex-direction: row;
-      flex-wrap: wrap;
-      gap: 12px;
-      background: none;
-      border: none;
-    }
-
-    .int-msg-answer {
-      border: 1px solid var(--customer-multi-color, #0e9dde);
-      color: var(--customer-multi-color, #0e9dde);
-      padding: 6px 14px;
-      border-radius: 8px;
-      cursor: pointer;
-      margin-bottom: 10px;
-    }
-
-    .int-msg-answer:hover {
-      background-color: var(--customer-multi-color, #0e9dde);
-      color: var(--customer-multi-hover-color, #fff)
-    }
-
-    @keyframes ellipsis {
-      to {
-        width: 10px;
-      }
-    }
-  `;
 
   protected render(): unknown {
     return html`
@@ -271,7 +141,7 @@ export class PubSocket extends LitElement // eslint-disable-line @typescript-esl
       this.canReply = false;
     }
 
-    this.dispatchEvent(new CustomEvent('can.reply', { detail: this.canReply }))
+    this.dispatchEvent(new CustomEvent('can.reply', {detail: this.canReply}))
 
     return html`
       <li ?customer=${false} ?undelivered="${msg.undelivered}" action-type="${msg.actionType}">
@@ -359,7 +229,7 @@ export class PubSocket extends LitElement // eslint-disable-line @typescript-esl
       time: (new Date()).getTime(),
       content: message,
       actionType: '',
-      meta: { id: ulid() },
+      meta: {id: ulid()},
       author: '',
       undelivered: true,
       customerInitiated: true
@@ -392,7 +262,7 @@ export class PubSocket extends LitElement // eslint-disable-line @typescript-esl
     if (this.scrolled) {
       this.newMessage = true;
     }
-    this.dispatchEvent(new CustomEvent('message', { detail: msg }))
+    this.dispatchEvent(new CustomEvent('message', {detail: msg}))
   }
 
   _scrollFn() {
@@ -446,6 +316,7 @@ export class PubSocket extends LitElement // eslint-disable-line @typescript-esl
       self.connected = true;
       self.connectionFailed = false;
       self._failureTime = 0;
+      self._retryAttempts = 0;
     }
 
     s.onclose = function (e) {
@@ -458,14 +329,14 @@ export class PubSocket extends LitElement // eslint-disable-line @typescript-esl
         if (self._failureTime <= 0) {
           self._failureTime = (new Date()).getTime();
         }
-        // should we retry, or fail?
-        if ((new Date()).getTime() < self._failureTime + (self.retryTimeout * 1000)) {
-          // retry
-          setTimeout(self.open.bind(self), self.retryDelay * 1000);
-        } else {
-          self.connectionFailed = true;
-          s.close(e.code, e.reason);
-        }
+
+        // Keep Retrying with exponential backoff
+        const delay = Math.min(Math.pow(2, self._retryAttempts) * self.retryDelay, 20);
+        console.log('Reconnect in ' + delay + 's');
+        setTimeout(() => {
+          self._retryAttempts++;
+          self.open();
+        }, delay * 1000);
       }
     };
