+++ tsconfig.json
@@ -1,39 +1,29 @@
 {
   "compilerOptions": {
+    "target": "ES2017",
     "experimentalDecorators": true,
-    "useDefineForClassFields": false,
-    "types": [
-      "node",
-      "mocha",
-      "chai",
-      "sinon"
-    ],
-    "moduleResolution": "node",
-    "module": "esnext",
-    "allowJs": true,
-    "target": "es6",
-    "sourceMap": false,
-    "declaration": false,
-    "inlineSources": false,
-    "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-    "strictNullChecks": true,
-    "plugins": [
-      {
-        "name": "ts-lit-plugin",
-        "strict": true
-      }
+    "moduleResolution": "Node",
+    "lib": [
+      "DOM",
+      "DOM.Iterable",
+      "ESNext"
     ],
-    "typeRoots": [
-      "node_modules/@types",
-      "./src/types"
-    ]
+    "types": [
+      "vite/client",
+      "node"
+    ],
+    "resolveJsonModule": true
   },
+  "typeRoots": [
+    "node_modules/@types",
+    "./src/types"
+  ],
   "include": [
-    "./src/"
+    "src"
   ],
   "exclude": [
-    "./node_modules/",
-    "./builds/"
+    "node_modules",
+    "builds"
   ]
 }
